# contracts/openapi/gateway-v1.yaml
openapi: 3.1.0
info:
  title: ManPaSik Gateway API
  version: 1.0.0
  description: |
    만파식 플랫폼 REST API 계약 (Gateway → gRPC 마이크로서비스 프록시).
    모든 엔드포인트는 /api/v1/ 프리픽스. 인증은 Bearer JWT.
  contact:
    name: ManPaSik Platform
servers:
  - url: https://api.manpasik.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 회원가입
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 가입 성공
...
  /users/{userId}/profile:
    get:
      tags: [User]
      summary: 프로필 조회
      operationId: getProfile
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
    put:
      tags: [User]
      summary: 프로필 수정
      operationId: updateProfile

  # ===== Measurement =====
...
  /rewards/{userId}/balance:
    get:
      tags: [Rewards]
      summary: 사용자 리워드 잔액 조회
      operationId: getRewardBalance
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
  /rewards/{userId}/contributions:
    get:
      tags: [Rewards]
      summary: 사용자 데이터 기여 이력 조회
      operationId: getRewardContributions
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }

  # ===== GxP Compliance =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }